move stack navigator props to navigation interface file
