+++ thoughtbubble-mobile/src/navigation/AppTabsNavigator.tsx
@@ -5,7 +5,7 @@ import { SettingsScreen } from '../screens/SettingsScreen';
 import { ProjectsStackNavigator } from './ProjectsStackNavigator';
 import { StatsStackNavigator } from './StatsStackNavigator';
 import { TabsParamList } from '../interfaces/navigation';
-import { AppNavTabsProps } from '../interfaces/componentProps';
+import { AppNavTabsProps } from '../interfaces/navigation';
 import { darkMode, lightMode } from '../constants/colors';
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
 import { useDarkCheck } from '../hooks/useDarkCheck';
