+++ thoughtbubble-mobile/src/navigation/StatsStackNavigator.tsx
@@ -5,10 +5,11 @@ import { StatsProjectInfoScreen } from '../screens/StatsProjectInfoScreen';
 import { StatsStackParamList } from '../interfaces/navigation';
 import { darkMode, lightMode } from '../constants/colors';
 import { useDarkCheck } from '../hooks/useDarkCheck';
+import { StatsStackNavigatorProps } from '../interfaces/navigation';
 
 const Stack = createStackNavigator<StatsStackParamList>();
 
-export const StatsStackNavigator: FC = function () {
+export const StatsStackNavigator: FC<StatsStackNavigatorProps> = function () {
   const isDarkMode = useDarkCheck();
   const dynamicScreenOptions = {
     headerStyle: {
